requirements.txt
requests
beautifulsoup4
python-telegram-bot==13.15
selenium
